+++ src/AppRouter.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import { BrowserRouter, Route, Routes, Navigate } from "react-router-dom";
 
 import LoginScreen from "./screens/LoginScreen/LoginScreen";
+import KakaoLoginRedirectScreen from "./screens/KakaoLogin/KakaoLoginRedirectScreen";
 import HomeScreen from "./screens/HomeScreen/HomeScreen";
 
 function AppRouter() {
@@ -10,7 +11,7 @@ function AppRouter() {
             <BrowserRouter basename="/">
                 <Routes>
                     <Route path="/" element={<LoginScreen/>} />
-                    <Route path="auth/kakao" element={<HomeScreen/>} />
+                    <Route path="auth/kakao" element={<KakaoLoginRedirectScreen/>} />
                     <Route path="auth/naver" element={<HomeScreen/>} />
                 </Routes>
             </BrowserRouter>
